requirements.txt
numpy
Pillow
pytorch_lightning
PyYAML
PyYAML
scikit_learn
torch==1.10.1
torchmetrics
torchvision
matplotlib
